+++ docs/master/Concurrent/Promises/Event.html
@@ -357,12 +357,12 @@ resolves.</p>
       <pre class="lines">
 
 
-857
-858
-859</pre>
+842
+843
+844</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 857</span>
+      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 842</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_any'>any</span><span class='lparen'>(</span><span class='id identifier rubyid_event_or_future'>event_or_future</span><span class='rparen'>)</span>
   <span class='const'>AnyResolvedEventPromise</span><span class='period'>.</span><span class='id identifier rubyid_new_blocked_by2'>new_blocked_by2</span><span class='lparen'>(</span><span class='kw'>self</span><span class='comma'>,</span> <span class='id identifier rubyid_event_or_future'>event_or_future</span><span class='comma'>,</span> <span class='ivar'>@DefaultExecutor</span><span class='rparen'>)</span><span class='period'>.</span><span class='id identifier rubyid_event'>event</span>
@@ -411,13 +411,13 @@ In other words, it inserts delay into the chain of Futures making rest of it laz
       <pre class="lines">
 
 
-867
-868
-869
-870</pre>
+852
+853
+854
+855</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 867</span>
+      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 852</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_delay'>delay</span>
   <span class='id identifier rubyid_event'>event</span> <span class='op'>=</span> <span class='const'>DelayPromise</span><span class='period'>.</span><span class='id identifier rubyid_new'>new</span><span class='lparen'>(</span><span class='ivar'>@DefaultExecutor</span><span class='rparen'>)</span><span class='period'>.</span><span class='id identifier rubyid_event'>event</span>
@@ -488,15 +488,15 @@ delay into the chain.</p>
       <pre class="lines">
 
 
-879
-880
-881
-882
-883
-884</pre>
+864
+865
+866
+867
+868
+869</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 879</span>
+      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 864</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_schedule'>schedule</span><span class='lparen'>(</span><span class='id identifier rubyid_intended_time'>intended_time</span><span class='rparen'>)</span>
   <span class='id identifier rubyid_chain'>chain</span> <span class='kw'>do</span>
@@ -547,12 +547,12 @@ delay into the chain.</p>
       <pre class="lines">
 
 
-897
-898
-899</pre>
+882
+883
+884</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 897</span>
+      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 882</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_to_event'>to_event</span>
   <span class='kw'>self</span>
@@ -600,14 +600,14 @@ delay into the chain.</p>
       <pre class="lines">
 
 
-889
-890
-891
-892
-893</pre>
+874
+875
+876
+877
+878</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 889</span>
+      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 874</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_to_future'>to_future</span>
   <span class='id identifier rubyid_future'>future</span> <span class='op'>=</span> <span class='const'><span class='object_link'><a href="../Promises.html" title="Concurrent::Promises (module)">Promises</a></span></span><span class='period'>.</span><span class='id identifier rubyid_resolvable_future'><span class='object_link'><a href="FactoryMethods.html#resolvable_future-instance_method" title="Concurrent::Promises::FactoryMethods#resolvable_future (method)">resolvable_future</a></span></span>
@@ -658,12 +658,12 @@ Any futures depending on it will use the new default executor.</p>
       <pre class="lines">
 
 
-903
-904
-905</pre>
+888
+889
+890</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 903</span>
+      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 888</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_with_default_executor'>with_default_executor</span><span class='lparen'>(</span><span class='id identifier rubyid_executor'>executor</span><span class='rparen'>)</span>
   <span class='const'>EventWrapperPromise</span><span class='period'>.</span><span class='id identifier rubyid_new_blocked_by1'>new_blocked_by1</span><span class='lparen'>(</span><span class='kw'>self</span><span class='comma'>,</span> <span class='id identifier rubyid_executor'>executor</span><span class='rparen'>)</span><span class='period'>.</span><span class='id identifier rubyid_event'>event</span>
@@ -719,16 +719,16 @@ then the result is as described in <span class='object_link'><a href="FactoryMet
       <pre class="lines">
 
 
-843
-844
-845
-846
-847
-848
-849</pre>
+828
+829
+830
+831
+832
+833
+834</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 843</span>
+      <pre class="code"><span class="info file"># File 'lib/concurrent/promises.rb', line 828</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_zip'>zip</span><span class='lparen'>(</span><span class='id identifier rubyid_other'>other</span><span class='rparen'>)</span>
   <span class='kw'>if</span> <span class='id identifier rubyid_other'>other</span><span class='period'>.</span><span class='id identifier rubyid_is_a?'>is_a?</span><span class='lparen'>(</span><span class='const'><span class='object_link'><a href="Future.html" title="Concurrent::Promises::Future (class)">Future</a></span></span><span class='rparen'>)</span>
