+++ docs/master/Concurrent/TimerSet.html
@@ -369,12 +369,12 @@ not running.</p>
       <pre class="lines">
 
 
-64
-65
-66</pre>
+62
+63
+64</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'lib/concurrent/executor/timer_set.rb', line 64</span>
+      <pre class="code"><span class="info file"># File 'lib/concurrent/executor/timer_set.rb', line 62</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_kill'>kill</span>
   <span class='id identifier rubyid_shutdown'>shutdown</span>
@@ -518,9 +518,7 @@ is successful; false after shutdown.</p>
 53
 54
 55
-56
-57
-58</pre>
+56</pre>
     </td>
     <td>
       <pre class="code"><span class="info file"># File 'lib/concurrent/executor/timer_set.rb', line 48</span>
@@ -528,11 +526,9 @@ is successful; false after shutdown.</p>
 <span class='kw'>def</span> <span class='id identifier rubyid_post'>post</span><span class='lparen'>(</span><span class='id identifier rubyid_delay'>delay</span><span class='comma'>,</span> <span class='op'>*</span><span class='id identifier rubyid_args'>args</span><span class='comma'>,</span> <span class='op'>&amp;</span><span class='id identifier rubyid_task'>task</span><span class='rparen'>)</span>
   <span class='id identifier rubyid_raise'>raise</span> <span class='const'>ArgumentError</span><span class='period'>.</span><span class='id identifier rubyid_new'>new</span><span class='lparen'>(</span><span class='tstring'><span class='tstring_beg'>&#39;</span><span class='tstring_content'>no block given</span><span class='tstring_end'>&#39;</span></span><span class='rparen'>)</span> <span class='kw'>unless</span> <span class='id identifier rubyid_block_given?'>block_given?</span>
   <span class='kw'>return</span> <span class='kw'>false</span> <span class='kw'>unless</span> <span class='id identifier rubyid_running?'>running?</span>
-  <span class='id identifier rubyid_opts'>opts</span> <span class='op'>=</span> <span class='lbrace'>{</span>
-    <span class='label'>executor:</span> <span class='ivar'>@task_executor</span><span class='comma'>,</span>
-    <span class='label'>args:</span> <span class='id identifier rubyid_args'>args</span><span class='comma'>,</span>
-    <span class='label'>timer_set:</span> <span class='kw'>self</span>
-  <span class='rbrace'>}</span>
+  <span class='id identifier rubyid_opts'>opts</span> <span class='op'>=</span> <span class='lbrace'>{</span> <span class='label'>executor:</span>  <span class='ivar'>@task_executor</span><span class='comma'>,</span>
+           <span class='label'>args:</span>      <span class='id identifier rubyid_args'>args</span><span class='comma'>,</span>
+           <span class='label'>timer_set:</span> <span class='kw'>self</span> <span class='rbrace'>}</span>
   <span class='id identifier rubyid_task'>task</span> <span class='op'>=</span> <span class='const'><span class='object_link'><a href="ScheduledTask.html" title="Concurrent::ScheduledTask (class)">ScheduledTask</a></span></span><span class='period'>.</span><span class='id identifier rubyid_execute'><span class='object_link'><a href="ScheduledTask.html#execute-class_method" title="Concurrent::ScheduledTask.execute (method)">execute</a></span></span><span class='lparen'>(</span><span class='id identifier rubyid_delay'>delay</span><span class='comma'>,</span> <span class='id identifier rubyid_opts'>opts</span><span class='comma'>,</span> <span class='op'>&amp;</span><span class='id identifier rubyid_task'>task</span><span class='rparen'>)</span> <span class='comment'># may raise exception
 </span>  <span class='id identifier rubyid_task'>task</span><span class='period'>.</span><span class='id identifier rubyid_unscheduled?'>unscheduled?</span> <span class='op'>?</span> <span class='kw'>false</span> <span class='op'>:</span> <span class='id identifier rubyid_task'>task</span>
 <span class='kw'>end</span></pre>
