+++ docs/master/_index.html
@@ -580,13 +580,6 @@
                 
               </li>
             
-              <li>
-                <span class='object_link'><a href="Concurrent/ErlangActor/Exit.html" title="Concurrent::ErlangActor::Exit (class)">Exit</a></span>
-                
-                  <small>(Concurrent::ErlangActor)</small>
-                
-              </li>
-            
           </ul>
         </ul>
       
@@ -655,20 +648,6 @@
           <ul>
             
             
-            
-              <li>
-                <span class='object_link'><a href="Concurrent/ErlangActor/HasFrom.html" title="Concurrent::ErlangActor::HasFrom (module)">HasFrom</a></span>
-                
-                  <small>(Concurrent::ErlangActor)</small>
-                
-              </li>
-            
-              <li>
-                <span class='object_link'><a href="Concurrent/ErlangActor/HasReference.html" title="Concurrent::ErlangActor::HasReference (module)">HasReference</a></span>
-                
-                  <small>(Concurrent::ErlangActor)</small>
-                
-              </li>
             
               <li>
                 <span class='object_link'><a href="Concurrent/Hash.html" title="Concurrent::Hash (class)">Hash</a></span>
@@ -1221,6 +1200,13 @@
                 <span class='object_link'><a href="Concurrent/Channel/Selector/TakeClause.html" title="Concurrent::Channel::Selector::TakeClause (class)">TakeClause</a></span>
                 
                   <small>(Concurrent::Channel::Selector)</small>
+                
+              </li>
+            
+              <li>
+                <span class='object_link'><a href="Concurrent/ErlangActor/Terminated.html" title="Concurrent::ErlangActor::Terminated (class)">Terminated</a></span>
+                
+                  <small>(Concurrent::ErlangActor)</small>
                 
               </li>
